Rewire the museum-hall gallery to the story-derived example layer
GalleryPage/galleryRender came from main built against the deleted
registry (new Function over gf-style code strings). The wall now sources
the 64 gallery-tagged stories via storyExamples.data.js and renders
frames by executing the real story modules: storyRender.ts extracts the
shared resolve/render surface (glob, loaders, await render) used by both
GoFishExample and the gallery; measure() is async and mounts before
rendering so getBBox-based layout works. Museum design, lazy/velocity
gating, search, and plaque links to the generated example pages are
unchanged.
